Update typo in graphic
The graphic contains text "Authenticate with Client ID + Client Secret to /token". Should this be "Authenticate with Client ID + Client Secret / token"? The 'to' seems like a typo. I can't submit this change without a change to the text so I've made a whitespace change above but please ignore it.